feat: implement git --follow -- <file_patn>
